Refactor+Feat: Restructured layout to allow scw deployment + implement deployment script
