+++ .github/workflows/test-on-pr.yml
@@ -1,0 +1,33 @@
+name: Test on push
+on:
+    pull_request:
+        branches:
+            - main
+            - dev
+jobs: 
+ build:
+  runs-on: ubuntu-latest
+  steps:
+    - name: Checkout Code
+      uses: actions/checkout@v4
+    
+    - name: Setup Go 1.21.x
+      uses: actions/setup-go@v4
+      with:
+        go-version: '1.20'
+    
+    - name: Install all dependencies
+      run: cd ${{github.workspace}}/scw_secret_wrapper && go mod get . && cd ${{github.workspace}}/helper && go mod get . && cd ${{github.workspace}}/sign && go mod get . && cd ${{github.workspace}}/cronjob && go mod get .
+
+    - name: Sign tests
+      run: cd ${{github.workspace}}/sign && go test -v
+    
+    - name: Run scw_secret_wrapper tests
+      run: cd ${{github.workspace}}/scw_secret_wrapper && go test ./... -v
+    
+    - name: Run helper tests
+      run: cd ${{github.workspace}}/helper && go test -v
+    
+    - name: Run cronjob tests
+      run: cd ${{github.workspace}}/cronjob && go test -v
+
