+++ .gitignore
@@ -20,3 +20,6 @@ node_modules/
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
+
+# ignore vendor folder
+vendor/
