+++ go.work
@@ -1,7 +1,10 @@
 go 1.20
 
 use (
-	./backend
 	./cronjob
 	./scw_secret_wrapper
+	./get_new_key
+	./has_new_key
+	./sign
+	./helper
 )
