+++ helper/api_key_retriver.go
@@ -1,4 +1,4 @@
-package decentproof_functions
+package helper
 
 import (
 	"os"
