+++ scw_secret_wrapper/go.mod
@@ -3,7 +3,8 @@ module github.com/Flajt/decentproof-backend/scw_secret_wrapper
 go 1.20
 
 require (
-	github.com/joho/godotenv v1.5.1 // indirect
-	github.com/scaleway/scaleway-sdk-go v1.0.0-beta.21 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
+	github.com/joho/godotenv v1.5.1
+	github.com/scaleway/scaleway-sdk-go v1.0.0-beta.21
 )
+
+require gopkg.in/yaml.v2 v2.4.0 // indirect
