+++ serverless.yml
@@ -1,0 +1,64 @@
+service: functions
+configValidationMode: off
+singleSource: false
+
+useDotenv: true
+
+provider:
+  name: scaleway
+  runtime: go120
+  env:
+    DEBUG: "false"
+  #secret:
+  #  PROJECT_KEY: ${env:PROJECT_KEY}
+  
+  scwToken: ${env:SCW_TOKEN}
+  scwProject: ${env:SCW_PROJECT}
+  # region in which the deployment will happen (default: fr-par)
+  scwRegion: nl-ams
+
+plugins:
+  - serverless-scaleway-functions
+
+package:
+  patterns:
+    - '!.gitignore'
+    - '!.git/**'
+    - '!*.sh'
+    - 'node_modules/**'
+    - '!.env'
+    - '!tsconfig.json'
+    - '!test/**'
+    - '!debug-server'
+    - '!build/debug-server'
+    - '!build/test/**'
+    - '!legacy/**'
+    - '!go.work'
+    - '!go.work.sum'
+    - '!LICENSE'
+    - '!README.md'
+    - '!util'
+    - 'cronjob/**'
+    - '!tests/**'
+
+functions:
+  new-key-required:
+    handler: has_new_key/HandleHasNewKey
+    httpOption: redirected
+
+  sign:
+    handler: sign/HandleSignature
+    httpOption: redirected
+
+  get-new-key:
+    handler: get_new_key/HandleGetNewKey
+    httpOption: redirected
+
+  cron-job:
+    handler: cronjob/Handle
+    events:
+      - schedule:
+        # Every Sunday at 00:00 => https://crontab.guru/#0_0_*_*_0
+        rate: "0 0 * * 0" # minute, hour, day (month) , month, day (week)
+        input:
+          key: "value" # is apperantly required....
